requirements.txt
click==7.0
requests==2.20.0
dateutils==0.6.6
logbook==1.4.0
pandas==0.24.1
aiohttp==3.5.4
aiodns==1.1.1
cchardet==2.1.4
async-timeout==3.0.1
dataclasses==0.6
lazy-object-proxy==1.3.1
numpy>=1.8.1
numexpr>=2.5.2
tables==3.4.4
TA-Lib